1. 样品进度查询功能基本实现，可以查看当前用户总体的申请情况，也可以查看某一申请的具体内容（申请的样品，申请时间等） 2. 样品申请详细内容查看界面设计完毕

+++ ESBtest/ViewModel/SampleRequestViewModel.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
+using System.Windows.Controls;
 
 namespace ESBtest.ViewModel
 {
@@ -28,8 +29,12 @@ namespace ESBtest.ViewModel
         public CommandBase ToolSampleRequestTab { get; set; }
         public CommandBase ToolProgressTab { get; set; }
         public CommandBase ToolHistoryTab { get; set; }
+        public CommandBase ToolApprovalTab { get; set; }
 
         public CommandBase UploadRequestCommand { get; set; }
+
+        public CommandBase CheckDetailsCommand { get; set; }
+        public CommandBase CancelRequestCommand { get; set; }
 
 
 
@@ -81,20 +86,46 @@ namespace ESBtest.ViewModel
             this.ToolSampleRequestTab = new CommandBase();
             this.ToolProgressTab = new CommandBase();
             this.ToolHistoryTab = new CommandBase();
+            this.ToolApprovalTab = new CommandBase();
             this.UploadRequestCommand = new CommandBase();
+            this.CheckDetailsCommand = new CommandBase();
+            this.CancelRequestCommand = new CommandBase();
 
-            //提交申请命令
-            this.UploadRequestCommand.ExecuteAction = new Action<object>(UploadRequest);
             //转到样品申请界面命令
             this.ToolSampleRequestTab.ExecuteAction = new Action<object>(SampleRequsetTab);
             //转到进度查询界面命令
             this.ToolProgressTab.ExecuteAction = new Action<object>(ProgressTab);
             //转到历史记录界面命令
             this.ToolHistoryTab.ExecuteAction = new Action<object>(HistoryTab);
+            //转到审批界面命令
+            this.ToolApprovalTab.ExecuteAction = new Action<object>(ApprovalTab);
+            //提交申请命令
+            this.UploadRequestCommand.ExecuteAction = new Action<object>(UploadRequest);
+            //查看申请细节命令
+            this.CheckDetailsCommand.ExecuteAction = new Action<object>(CheckDetails);
+            //取消申请命令
+            this.CancelRequestCommand.ExecuteAction = new Action<object>(CancelRequest);
             #endregion
         }
 
         #region 功能命令实现
+        /// <summary>
+        /// 刷新样品申请记录
+        /// </summary>
+        /// <param name="dg"></param>
+        private void RefreshSampleRecordDataGrid(DataGrid dg, int flag)
+        {
+            dg.ItemsSource = null;
+            if(flag == 1)
+            {
+                this.SampleRecordModelList = dBControl.SearchRecord(GlobalValue.CurrentUser.UserID);
+            }
+            else if (flag == 2)
+            {
+
+            }
+            dg.ItemsSource = this.SampleRecordModelList;
+        }
         /// <summary>
         /// 切换至样品申请界面
         /// </summary>
@@ -110,7 +141,7 @@ namespace ESBtest.ViewModel
         private void ProgressTab(object w)
         {
             (w as SampleRequestView).TabControlFunction.SelectedIndex = 1;
-
+            RefreshSampleRecordDataGrid((w as SampleRequestView).ProgressDataGrid, 1);
         }
         /// <summary>
         /// 切换至历史记录界面
@@ -121,16 +152,30 @@ namespace ESBtest.ViewModel
             (w as SampleRequestView).TabControlFunction.SelectedIndex = 2;
         }
         /// <summary>
+        /// 切换到审批界面
+        /// </summary>
+        /// <param name="w"></param>
+        private void ApprovalTab(object w)
+        {
+            (w as SampleRequestView).TabControlFunction.SelectedIndex = 3;
+
+        }
+        /// <summary>
         /// 提交样品申请
         /// </summary>
         /// <param name="w">SampleRequestView</param>
         private void UploadRequest(object w)
         {
             List<int> iList = dBControl.SearchInCart(GlobalValue.CurrentUser.UserID);
-            string requestDate = SampleRecord.RequestDate.ToShortDateString().ToString();
-            if(dBControl.InsertIntoSampleRecordTable(iList, requestDate, GlobalValue.CurrentUser.UserID, 1) > 0)
+            if (iList.Count <= 0)
             {
-                if(dBControl.DeleteCartTable(GlobalValue.CurrentUser.UserID) > 0 )
+                MessageBox.Show((w as Window), "当前购物车中没有样品", "提示");
+                return;
+            }
+            string requestDate = SampleRecord.RequestDate.ToShortDateString().ToString();
+            if (dBControl.InsertIntoSampleRecordTable(iList, requestDate, GlobalValue.CurrentUser.UserID, 1) > 0)
+            {
+                if (dBControl.DeleteCartTable(GlobalValue.CurrentUser.UserID) > 0)
                 {
                     MessageBox.Show((w as Window), "申请提交成功", "提示");
                 }
@@ -142,6 +187,35 @@ namespace ESBtest.ViewModel
             else
             {
                 MessageBox.Show((w as Window), "申请提交失败", "提示");
+            }
+        }
+        /// <summary>
+        /// 查看选中的样品申请的详情
+        /// </summary>
+        /// <param name="w"></param>
+        private void CheckDetails(object w)
+        {
+            SampleRecordDetailView srdv = new SampleRecordDetailView();
+            srdv.ControllerGrid.Visibility = Visibility.Collapsed;
+            List<int> iList = dBControl.SearchSampleRecord((w as SampleRecordModel).IdRecord, GlobalValue.CurrentUser.UserID);
+            (srdv.DataContext as SampleRecordDetailViewModel).SampleModelList = dBControl.SearchSample(iList);
+            (srdv.DataContext as SampleRecordDetailViewModel).SampleRecordModel = dBControl.SearchRecord((w as SampleRecordModel).IdRecord, GlobalValue.CurrentUser.UserID);
+            srdv.ShowDialog();
+        }
+        /// <summary>
+        /// 取消选中的样品申请
+        /// </summary>
+        /// <param name="w"></param>
+        private void CancelRequest(object w)
+        {
+            SampleRecordModel sr = (SampleRecordModel)(w as DataGrid).SelectedItem;
+            if(MessageBox.Show("确认取消此条申请记录?","提示", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+            {
+                if(dBControl.UpdateRecordTable(sr.IdRecord, GlobalValue.CurrentUser.UserID, DateTime.Now, DateTime.Now, DateTime.Now, 10) > 0)
+                {
+                    MessageBox.Show("取消成功", "提示");
+                    RefreshSampleRecordDataGrid((w as SampleRequestView).ProgressDataGrid);
+                }
             }
         }
         #endregion
